test(racingcar): ResultTest의 메서드를 GameCarsTest로 이동
